Move all wallet functionality to separate dir
