+++ web/src/wallet/Wallet.tsx
@@ -4,6 +4,11 @@ import { makeStyles } from '@material-ui/core/styles';
 import { Button } from '@material-ui/core';
 import Paper from '@material-ui/core/Paper';
 import Typography from '@material-ui/core/Typography';
+import Config from '../config';
+import localWallet from './localWallet';
+import keplr from './keplrWallet';
+
+const config = Config();
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -21,13 +26,18 @@ const useStyles = makeStyles((theme) => ({
 
 interface Props {
   client: SecretJS.SigningCosmWasmClient | undefined;
+  setClient: (client: SecretJS.SigningCosmWasmClient) => void;
   faucetUrl: string | undefined;
 }
 
 export default (props: Props) => {
   const classes = useStyles();
-  const { client, faucetUrl } = props;
+  const { client, setClient, faucetUrl } = props;
 
+  useEffect(() => {
+    //localWallet(config.lcdUrl, setClient);
+    keplr(config.chainId, config.chainName, config.lcdUrl, config.rpcUrl, setClient);
+  }, []);
   const [account, setAccount] = useState<SecretJS.Account | undefined>();
   useEffect(() => {
     if (!client) return;
